requirements.txt
Flask
mariadb
python-dotenv